refactor to use provider method
